fix: remove the slash from index option
I'm trying to set my index to something else in my webpack config.
according the README and the code here:

https://github.com/webpack/webpack-dev-middleware#index
https://github.com/webpack/webpack-dev-middleware/blob/master/lib/middleware.js#L81

I changed my config here.
```
// config.js
{ index: 'otherpage.html' }
// dev-server.js
webpackDevMiddleware({ index: config.index })
```
but it doesn't work as I expected. I found out `connect-history-api-fallback` will rewrite the url to 'index.html' from here:
https://github.com/bripkens/connect-history-api-fallback/blob/master/lib/index.js#L82

so I have to add  the following code to fix it:

```
history({
  index: config.index
});
```
but actually this is wrong, because I'm missing a '/' for it.

so I would expect the `index` always be the same meaning.